Move the spiner param struct and some helpers into singularity-utils
